Updated commenting with EVC7 class

+++ Testcase/Telegrams/EVCtoDMI/EVC7_MMIEtcsMiscOutSignals.cs
@@ -2,12 +2,23 @@
 
 namespace Testcase.Telegrams.EVCtoDMI
 {
+    /// <summary>
+    /// The signals in this telegram are outputs from the generic ETCS OB R4 system.
+    /// This telegram collects miscellaneous output signals to the train.
+    /// Note:
+    /// This packet is routed via dedicated port and thus no header nor length information is contained in the (plain) data set.
+    /// It is also protected via SDTv2.
+    /// </summary>
     static class EVC7_MMIEtcsMiscOutSignals
     {
         private static SignalPool _pool;
         private static MMI_OBU_TR_BRAKETEST_STATUS _brakeTestStatus;
         private static MMI_OBU_TR_M_LEVEL _level;
 
+        /// <summary>
+        /// Initialise EVC-7 telegram
+        /// </summary>
+        /// <param name="pool"></param>
         public static void Initialise(SignalPool pool)
         {
             _pool = pool;
@@ -33,6 +44,17 @@ namespace Testcase.Telegrams.EVCtoDMI
             _pool.SITR.ETCS1.EtcsMiscOutSignals.EVC7alias1B1.Value = (byte) (brakeTestStatus << 4 | level);
         }
 
+        /// <summary>
+        /// EB test in progress
+        /// 
+        /// Values
+        /// 0 = BrakeTestNotInProgress
+        /// 1 = BrakeTestInProgress
+        /// 2 = BrakeTestSuccessful
+        /// 3 = BrakeTestFailed
+        /// 4 = UnableToStartBrakeTest
+        /// 5 = Aborted
+        /// </summary>
         public static MMI_OBU_TR_BRAKETEST_STATUS MMI_OBU_TR_BrakeTest_Status
         {
             set
@@ -42,6 +64,18 @@ namespace Testcase.Telegrams.EVCtoDMI
             }
         }
 
+        /// <summary>
+        /// ETCS Mode
+        /// 
+        /// Values:
+        /// 0 = "Level 0"
+        /// 1 = "Level NTC"
+        /// 2 = "Level 1"
+        /// 3 = "Level 2"
+        /// 4 = "Level 3"
+        /// 5..254 = "spare"
+        /// 255 = "unknown"
+        /// </summary>
         public static MMI_OBU_TR_M_LEVEL MMI_OBU_TR_M_Level
         {
             set
@@ -51,12 +85,40 @@ namespace Testcase.Telegrams.EVCtoDMI
             }
         }
 
+        /// <summary>
+        /// ETCS Mode
+        /// 
+        /// Values:
+        /// 0 = "FS - Full Supervision"
+        /// 1 = "OS - On-sight"
+        /// 2 = "SR - Staff Responsible"
+        /// 3 = "SH - Shunting"
+        /// 4 = "UN - Unfitted"
+        /// 5 = "SL - Sleeping"
+        /// 6 = "SB - Standby"
+        /// 7 = "TR - Trip"
+        /// 8 = "PT - Post trip"
+        /// 9 = "SF - System failure"
+        /// 10 = "IS - Isolation"
+        /// 11 = "NL - Non-leading"
+        /// 12 = "LS - Limited Supervision"
+        /// 13 = "SN - National System"
+        /// 14 = "RV - Reversing"
+        /// 15 = "PS - Passive Shunting"
+        /// 16 = "NP - No Power"
+        /// 17..127 = "Not used"
+        /// 128 = "Unknown" (DEFAULT)
+        /// 129..255 = "Not used"
+        /// </summary>
         public static MMI_OBU_TR_M_MODE MMI_OBU_TR_M_Mode
         {
             set => _pool.SITR.ETCS1.EtcsMiscOutSignals.MmiObuTrMMode.Value = (byte) value;
         }
 
-        public enum MMI_OBU_TR_BRAKETEST_STATUS : ushort
+        /// <summary>
+        /// Brake test mode enum
+        /// </summary>
+        public enum MMI_OBU_TR_BRAKETEST_STATUS : byte
         {
             BrakeTestNotInProgress = 0,
             BrakeTestInProgress = 1,
@@ -66,7 +128,10 @@ namespace Testcase.Telegrams.EVCtoDMI
             Aborted = 5
         }
 
-        public enum MMI_OBU_TR_M_LEVEL : ushort
+        /// <summary>
+        /// ETCS Level enum
+        /// </summary>
+        public enum MMI_OBU_TR_M_LEVEL : byte
         {
             L0 = 0,
             LNTC = 1,
@@ -76,7 +141,10 @@ namespace Testcase.Telegrams.EVCtoDMI
             Unknown = 15
         }
 
-        public enum MMI_OBU_TR_M_MODE : ushort
+        /// <summary>
+        /// ETCS Mode enum
+        /// </summary>
+        public enum MMI_OBU_TR_M_MODE : byte
         {
             FullSupervision = 0,
             OnSight = 1,
